chore: move client types into .d.ts
This avoids duplication between src/client and src/node
