+++ packages/react-pages/src/client/SSRContextProvider.tsx
@@ -3,7 +3,7 @@
 import React, { useState } from 'react'
 import { HashRouter, BrowserRouter } from 'react-router-dom'
 import { dataCacheCtx, setDataCacheCtx } from './ssr/ctx'
-import type { PageLoaded } from './types'
+import type { PageLoaded } from '../../client'
 
 // @ts-expect-error
 const Router = __HASH_ROUTER__ ? HashRouter : BrowserRouter
