+++ packages/react-pages/src/client/index.tsx
@@ -1,1 +1,1 @@
-export type { Theme } from './types'
+export type { Theme } from '../../client'
