+++ packages/react-pages/src/client/ssr/serverRender.tsx
@@ -5,7 +5,7 @@ import { StaticRouter } from 'react-router-dom'
 import ssrData from '@!virtual-modules/ssrData'
 import App from '../App'
 import { dataCacheCtx } from './ctx'
-import type { PagesLoaded } from '../types'
+import type { PagesLoaded } from '../../../client'
 
 // put all page data in cache, so that we don't need to load it in ssr
 const dataCache: PagesLoaded = ssrData
