+++ packages/react-pages/src/node/index.ts
@@ -96,7 +96,8 @@ export default function pluginFactory(
 }
 
 export type {
-  Theme as Theme,
-  PagesStaticData as PagesStaticData,
-  PagesLoaded as PagesLoaded,
-} from './types/client'
+  Theme,
+  LoadState,
+  PagesLoaded,
+  PagesStaticData,
+} from '../../client'
